Implement the direct message page
added the Direct Messages Page as per the design.

made changes to the app.css files accordingly to style the page.

made the TextMessage.cs file as a model for the text messages to be typed in the chat box, this can be later on used for the main page chat box as well when implemented,