Dm page addition and style fixes
DM page has been created and the required stylings have been applied, send message button has been made but it's still not functional, but this is not within the scope of the current ticket at the moment.

different shading has also been applied to the message area as requested in a comment on the previous pull request.